Implement smarter processing of calibration points for the Bowling et al. 2003 carbon isotope calibration.
Added a skeleton to allow for points to be filtered out based on a set of criteria, and perform calibration if only 2 of the calibration data points are acceptable. Currently, only checks to see if expected [CO2] is w/in 10 ppm of measured [CO2]. This will help filter out a certain set of failure modes - namely, empty calibration tanks - but more work is likely necessary here.