Changed conc_thres from 10 to 25
Looking through BART data, and there are seemingly okay validation points that are thrown out because difference is too high - might need to revisit this though. Not sure I know about how accurate the concentration measurements on these instruments are, might be a good idea to compare to the IRGA some point later in time.