Fix bug causing all output d13C values to be NaN.
Prior tweaks had caused all output d13C values to be NANs, as no gain/offset values were calculated. This update corrects that change, and allows for dta to be properly calculated again. However - the uncertainty equations or inputs are almost certainly wrong, as the uncertainty values are unreasonably high.