Increase consistency between input and output files.
This is a major change, with variable names changing! A number of issues with trying to use neonUtilities on these files, which assume a very particular file structure, has convinced me it is worth remaining close to the NEON file/data structure. This commit makes these changes, but *it may break plotting scripts meant to work on versions of data I have produced to this point since variable names change*. If your scripts break, please check the variable names first.