Enforce all time differences to be in seconds.
Oooh...this was a fun one. diff on POSIXct can return different units (not always seconds!). As a result, sometimes differences were in seconds (and therefore, easy to ID calibration periods where the difference was >1800 seconds) - but sometimes they were in minutes (where, if measurements are made every day, NO difference should be >1800 minutes). Commit ensures that seconds are *always* returned. Closes #48.