Add options to plot variance in co2/d13c measurements in carbon_diagnostic_plots.
This is useful for diagnosing timestamp issues in NEON EC data - there is an existing issue in their data where timestamps of different instruments become out of sync, causing inappropriate aliasing of measurements across periods. This cannot be corrected in the L1 data, and must be corrected by NEON. Unfortunately, it affects up to ~50% of the archive.