remove if statement from top of correct_carbon_ref_cval.
Sites that had corrections were enumerated separately at the top in an if statement, but this risked skipping any new corrections added if they were added to the else branch, but the top of the funciton was not updated. The if statement only saves a miniscule amount of compute time across the whole calibration routine, so safer to remove it.